nfp-backend/pyproject.toml
[tool.poetry]
name = "nfp-backend"
version = "0.1.0"
description = ""
authors = ["Joe Tristano <joe.tristano+gh@gmail.com>"]
readme = "README.md"
packages = [{include = "nfp_backend"}]

[tool.poetry.dependencies]
python = "3.9.6"
alembic = "1.7.5"
anyio = "3.4.0"
asgiref = "3.4.1"
asyncpg = "0.25.0"
bcrypt = "3.2.0"
cffi = "1.15.0"
click = "8.0.3"
cryptography = "36.0.1"
databases = "0.5.3"
ecdsa = "0.17.0"
fastapi = "0.70.0"
greenlet = "1.1.2"
gunicorn = "20.1.0"
h11 = "0.12.0"
httptools = "0.2.0"
idna = "3.3"
mako = "1.1.6"
markupsafe = "2.0.1"
passlib = "1.7.4"
pyasn1 = "0.4.8"
pycparser = "2.21"
pydantic = "1.8.2"
python-dotenv = "0.19.2"
python-jose = "3.3.0"
python-multipart = "0.0.5"
pyyaml = "6.0"
rsa = "4.8"
six = "1.16.0"
sniffio = "1.2.0"
sqlalchemy = "1.4.27"
starlette = "0.16.0"
typing-extensions = "4.0.1"
uvicorn = "0.15.0"
uvloop = "0.16.0"
watchgod = "0.7"
websockets = "10.1"
psycopg2 = "^2.9.5"
pandas = "^1.5.3"
scipy = "^1.10.1"
jira = "^3.4.1"
pychart-js = "^0.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
